fix cast warning

+++ omsApp/src/drvMAXv.cc
@@ -1111,7 +1111,7 @@ static int motorIsrSetup(int card)
 #else
         (void (*)(void *)) motorIsr,
 #endif
-        (void *) card);
+        (void *)(size_t) card);
 
     if (!RTN_SUCCESS(status))
     {
