+++ omsApp/src/drvOms.cc
@@ -926,7 +926,7 @@ static int motorIsrEnable(int card)
 #else
                                                           (void (*)(void *)) motorIsr,
 #endif
-                                                          (void *) card);
+                                                          (void *)(size_t) card);
 
         if (!RTN_SUCCESS(status))
         {
